Removed JWT file and moved Roles to a file
