+++ src/common/seed.py
@@ -2,7 +2,8 @@
 # -*- coding: utf-8 -*-
 from model.features import add_features
 from model.attendee import add_attendee
-from model.user import add_user, add_roles
+from model.user import add_user
+from model.role import add_roles
 from model.event import add_event
 from loguru import logger
 import base64
@@ -46,14 +47,15 @@ def seed_user():
     else:
         logger.error('Failed to add user')
 
+
 def seed_roles():
-    data = [{'name':'Admin'}, {'name':'User'}]
+    data = [{'name': 'Admin'}, {'name': 'User'}]
     for row in data:
         result = add_roles(row)
         if result:
-            logger.success('Role Sucessfully added')
+            logger.success('{} role ucessfully added', row['name'])
         else:
-            logger.error('Failed to add Role')
+            logger.error('Failed to add {} role', row['name'])
 
 
 def seed_event():
