+++ src/model/user.py
@@ -1,5 +1,4 @@
 import os
-import jwt
 import bcrypt
 from loguru import logger
 from marshmallow_sqlalchemy import ModelSchema
@@ -7,11 +6,11 @@ from flask import jsonify
 from dotenv import load_dotenv
 from datetime import datetime, timedelta
 from flask_jwt_extended import (create_access_token)
+from flask_sqlalchemy import SQLAlchemy
 
 from .features import add_features, generate_features
+from .role import role_schema, get_user_roles, Role
 from common.common import gen_hash, db
-
-from flask_sqlalchemy import SQLAlchemy
 
 session = db.session
 
@@ -26,13 +25,6 @@ class User(db.Model):
     roles = db.relationship('Role',
                             secondary='user_roles',
                             backref=db.backref('users', lazy='joined'))
-
-
-# Define the Role data-model
-class Role(db.Model):
-    __tablename__ = 'roles'
-    id = db.Column(db.BIGINT(), primary_key=True)
-    name = db.Column(db.String(50), unique=True)
 
 
 # Define the UserRoles association table
@@ -52,15 +44,6 @@ class UserSchema(ModelSchema):
 
 user_schema = UserSchema()
 user_schemas = UserSchema(many=True)
-
-
-class RoleSchema(ModelSchema):
-    class Meta:
-        model = Role
-
-
-role_schema = RoleSchema()
-role_schemas = RoleSchema(many=True)
 
 
 class UserRoleSchema(ModelSchema):
@@ -104,41 +87,11 @@ def add_user(data):
         return didSucceed
 
 
-def add_roles(data):
-    didSucceed = False
-    role = Role(name=data['name'])
-    session.add(role)
-    try:
-        session.commit()
-        logger.debug("Role successfully added")
-        didSucceed = True
-    except Exception as e:
-        print("Error ==> ", e)
-        session.rollback()
-        raise
-    finally:
-        session.close()
-        return didSucceed
-
-
 def get_user(email):
     logger.info("Attempting to get user")
     try:
         user = session.query(User).filter_by(email=email).first()
         return user
-    except Exception as e:
-        print(e)
-
-
-def get_user_roles(user_id):
-    try:
-        user_data = User.query.join(UserRoles).join(Role).filter(
-            (UserRoles.user_id == user_id)
-            & (UserRoles.role_id == Role.id)).first()
-        role = None
-        for id in user_data.roles:
-            role = role_schema.dump(id)
-        return role['name']
     except Exception as e:
         print(e)
